Refactored classes as cleanup
Got a AspNetCore MVC going working with the TCP Listener so that I can do the SRP encryption stuff.
HTTP in dotnet does not let you easily capture incomming TCP connection on a HTTP Context... to say decrypt it first so had to do this proxy thing.

Got Homekit on iPhone going through TCP Listener and TCP Client proxy to MVC API get hit with the Content as Bytes!

That was the easy part? Phew... not that easy..